feat(core): com.malinskiy.marathon.execution tests migrated to junit
